requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Pillow == 9.5.0
dgl == 1.1
extra_streamlit_components == 0.1.56
networkx == 2.8.4
numpy == 1.24.3
pandas == 1.5.3
plotly == 5.14.1
rdkit_pypi == 2022.9.5
scikit_learn == 1.2.2
streamlit == 1.21.0
torch == 2.0.0
